@NotBuer -> Form cadastro compra gridview dynamic loading filtering

+++ Pecuaria/Models/AnimalCompraGadoItemOutput.cs
@@ -9,7 +9,7 @@ namespace Pecuaria.Models
     public class AnimalCompraGadoItemOutput
     {
         public string NomeAnimal { get; set; }
-        public short Quantidade { get; set; }
+        public int Quantidade { get; set; }
         public decimal Preco { get; set; }
         public decimal ValorTotal { get; set; }
     }
